🧪🚀 CI/CD Pipeline with GitHub Actions. #131
remove testing.yml because it is not needed anymore and add cicd.yml to run the tests and if they pass, deploy the app to GitHub Pages.
I updated the README.md file to explain how anyone can use this repo.
I also updated the vite.config.ts file and I add new base url for the app. Because GitHub Pages will be used to deploy the app, the base url will be the repo name.