🐛 fix(TEST Backend create Mock)
I removed all tests automatically generated when someone creates a new module. Because they are useless and it generates bugs!
I'm also creating a mock for the database, so we can test controllers and services without the database!
I also fixed removing all unused imports and variables.
I also added a new .env.example file, so we can create an .env file with our own configuration.
I also update the dependencies of the project because there were some vulnerabilities in the old ones.