✨ feat(add-user-excel) : create the view and the service to make a request to the API #132
I also add a new route to the router to access to the view.
And I add a new type to the user type. The type is for the excel file.
Finally, I update the package.json because I add a new package to read the excel file and update some packages to fix some vulnerabilities.