:sparkles: feat(header): Reworking the header:
We decided to add a sub header for the tab "Ajouter" because the 3 tabs => "Ajouter classes", "Ajouter Élève/prof" and "Ajouter Activité" were taking to much space in the header bar.
So we decided that when you hover the "Ajouter" tab it displays a "subtab" with the 3 categories in it.

Fixing the "Suivi" tab. It was displaying for all the users roles. Now only admin and teachers can see it.