feat(authentification): Cookie & JWT user authentification
Before we store the user information we receive during user login in
local storage. This allowed us to manipulate them as we pleased,
however, this involves a major security breach since we have access to
sensitive user information and a potential attacker could steal it.

First, local storage will no longer be used during user connections.
During the user connection, a JWT is generated which contains the
identifier of the user in the database. This JWT is introduced in the
cookie valid over a period of 24 hours contained in the response that is
sent back to the user. The advantage of using a cookie is that it is not
directly accessible by the js client and will be automatically returned
in the request header on each user request.
Secondly, when the user wishes to access a sensitive resource when he is
connected, an x-xsrf-token string is also sent in the request header,
which identifies the user's session. So after checking the x-xsrf-token
and the token, we get the information that the user needs.

issue #159