feat(route access): block route for non authentificated user
Before, the routes were exposed to all users, which allowed access from
the url to some administrator pages or to those who had no role!! which
poses a real security concern for the application

Now any user who is not authenticated will be redirected to the login
page even if he modifies the url of the browser. For this we perform a
first check to find out if the user is connected and then we
recover his role from the database and we check if he has the role to
access the page otherwise he is redirected to the login.

Another change was user registration. We perform a check on the user who
adds a student (his role and if he is authenticated)

BREAKING CHANGE: routes are no longer directly accessible without user
login

close #159 #166