feat(userResponse): adapt entities that we receive in user response table
When the user finish the card game, his result must to be store in db in
the user response table. But the query cann't pass because the
proficiency table is not feed and in the front we send the bad user id
of the actual user we also need to store the level map game that the
user play.

OTHER BUG : We can't get the image store in database for the drag and
learn gam when the user access to the level drag and learn.
In other to fix this bug we have to make some route in the
app.controller that get the path of the image and adapt the path in
other to get the image depend on your os.

BREAKING CHANGE : update table user response, make route for get the
image on the server, And refactor some variable to respect camelCase
convention.

issue #170